requirements.txt
azure-functions==1.18.0
azure-cosmos==4.3.1
azure-identity==1.14.1
azure-storage-blob==12.14.1
python-dateutil==2.8.2
